add panning
